requirements.txt
streamlit
requests
matplotlib
smtplib
python-dotenv